Finished refactoring. World, CellData, etc. no longer have type variables.
Instead, Agent and Wumpus are quantified over their minds. Entity has two type variables (for the minds of the Agents/Wumpuses), but CellData and World only contain the existentially quantified SomeMind. Wumpuses (which store the world as their internal state) get around the problem of cyclically referring to themselves - i.e. a Wumpus storing the World, which has a Wumpus which stores the world... - by erasing the internal states of other agents and Wumpuses when they make a copy of the world for their own use.