Re-did helper functions in Memory.hs.
constructCell, constructEdge now return update functions instead of CellData/EdgeData values with undefined values. This has the advantage of enabling partial updates to stored memory, esp. in the case of AMVisualFree, which indicates the absence of information. Before, we'd have had no way of differentiating between the positive knowledge that a cell's entity should be set to Nothing and the negative knowledge that it should not be changed.