Added Discharged-field to AMPlannedAction to signifiy which actions have already been inserted as memories.
Made the semantics of adding memories consistent: an AMPlannedAction has to contain an existing memory index. If the action is non-discharged, the belief generator will add a new memory node as the child of the one given by this index.

Removed beliefGeneratorComponent and renamed generateBelief to beliefGeneratorComponent (they were actually duplicates). Renamed generateBelief' to generateBelief.