Decision maker: local data is now passed to ALL cells during per-cell evaluation.
Since enthusiasm w.r.t. plants depends on the agent's current health, which is local information, local messages are now passed to all cells by the planner which evaluates the emotional impact of each cell.
In addition, evaluateCell also inserts an AMYouAreHere message for the (0,0) cell (relative index). This allows the enthusiasm filter to detect that it should evaluate its health specially. Given that every cell gets the AMHaveHealth message anyway, we need a way to distinguish the (0,0) cell, so as to generate high enthusiasm for at and thereby induce hungry agents to eat.