Started fixing the snafu with the BG and M. Doesn't compile.
The BG and the memory are still tangled up. Memories are inserted twice because the BG as well as the memory call addMemory. On the other hand, memory doesn't atually simulate the consequences of actions (reconstructWorld doesn't do this; simulateStep does). So we have a problem: memory inserts the CURRENT memory state, not the future, predicted one.

In the future, memory will be responsible for simulating future states as well as saving them, but it won't insert new messages. The BG, on the other hand, which gets called immediately after memory, won't create new memories; it will only read existing memories (through PlannedAction- or RecallMemory-messages) and insert the resultant messages.