Changed excitement mechanics.
An activated node now sends <es>% of its excitement to its neighbor,
where <es> is the strength of the edge going from source to target.

Before, it sent <t>*<es>, where <t> was the threshold of the target
node. This was quite impractical, because the edge strength of one
edge between nodes S, T had to depend on the edge strengths of all
other edges going to T.

Also, added lots of traces.