Added some traces and log output.
Added the function AgentMind.getFilters for debugging.

Changed the castable instance VisualAgent Agent: the agent gets a dummyMind,
not an AgentMind. The cast is only performed by the BG and such reconstructed
agents are supposed to be dummies.

The idea of inferring other agents' emotions and disposition towards one, and
inferring their actions from that, is an interesting one, though.